bring many tests back
