+++ test/runtests.jl
@@ -4,9 +4,9 @@ using ACEpotentials, Test, LazyArtifacts
 
     @testset "Models" begin include("models/test_models.jl") end 
 
-
-    # @testset "Test silicon" begin include("test_silicon.jl") end
-    # @testset "Test recomputation of weights" begin include("test_recompw.jl") end
+    # fitting tests 
+    @testset "Test silicon" begin include("test_silicon.jl") end
+    @testset "Test recomputation of weights" begin include("test_recompw.jl") end
 
     # TODO: bring FIO back cf Issue #217
     # @testset "Test IO" begin include("test_io.jl")  end
@@ -16,18 +16,10 @@ using ACEpotentials, Test, LazyArtifacts
     # @testset "UF_ACE" begin include("test_uface.jl") end
 
     # weird stuff 
-    # @testset "Weird bugs" begin include("test_bugs.jl") end
+    @testset "Weird bugs" begin include("test_bugs.jl") end
 
     # ACE1 compatibility tests 
     # TODO: these tests need to be revived either by creating a JSON 
     #       of test data, or by updating ACE1/ACE1x/JuLIP to be compatible. 
     # @testset "ACE1 Compat" begin include("ace1/test_ace1_compat.jl"); end 
-
-    # outdated
-    # @testset "Read data" begin include("outdated/test_data.jl") end 
-    # @testset "Basis" begin include("outdated/test_basis.jl") end 
-    # @testset "Solver" begin include("outdated/test_solver.jl") end 
-    # @testset "Fit ACE" begin include("outdated/test_fit.jl") end 
-    # @testset "Read params" begin include("outdated/test_read_params.jl") end 
-    # @testset "Test ace_fit.jl script" begin include("outdated/test_ace_fit.jl") end
 end
