+++ test/test_silicon.jl
@@ -3,6 +3,7 @@
 ##
 
 using ACEpotentials
+using ACEbase.Testing: println_slim
 using ExtXYZ, AtomsBase
 using Distributed
 using LazyArtifacts
@@ -12,87 +13,51 @@ using ACEpotentials.Models: ACEPotential
 
 ## ----- setup -----
 
-@warn "test_silicon not fully converted yet."
-
 params = (elements = [:Si],
           Eref = [:Si => -158.54496821],
           rcut = 5.5,
           order = 3,
           totaldegree = 12)
 
-model1a = acemodel(; params...)
-model1b = acemodel(; params...)
-model2 = ACEPotential( ACE1compat.ace1_model(; params...) )
+model = ACEPotential( ACE1compat.ace1_model(; params...) )
 
-data1 = read_extxyz(artifact"Si_tiny_dataset" * "/Si_tiny.xyz")
-data2 = ExtXYZ.load(artifact"Si_tiny_dataset" * "/Si_tiny.xyz")
+data = ExtXYZ.load(artifact"Si_tiny_dataset" * "/Si_tiny.xyz")
 
 data_keys = [:energy_key => "dft_energy",
-             :force_key => "dft_force",
+             :force_key  => "dft_force",
              :virial_key => "dft_virial"]
+
 weights = Dict("default" => Dict("E"=>30.0, "F"=>1.0, "V"=>1.0),
-               "liq" => Dict("E"=>10.0, "F"=>0.66, "V"=>0.25))
+                   "liq" => Dict("E"=>10.0, "F"=>0.66, "V"=>0.25))
 
 ## ----- perform tests -----
 
-# TODO : bring a target error back? 
-
-# function test_rmse(rmse, expected, atol)
-#     for config in keys(rmse)
-#         # TODO and/or warning: can't iterate over rmse because it will have virial for isolated atom
-#         #for obs in keys(rmse[config])
-#         for obs in keys(expected[config])
-#             @test rmse[config][obs] ≈ expected[config][obs] atol=atol
-#         end
-#     end
-# end
-
-# # @testset "QR" begin
-# rmse_qr = Dict(
-#     "isolated_atom" => Dict("E"=>0.0, "F"=>0.0),
-#     "dia"           => Dict("V"=>0.0234649, "E"=>0.000617953, "F"=>0.018611),
-#     "liq"           => Dict("V"=>0.034633, "E"=>0.000133371, "F"=>0.104112),
-#     "set"           => Dict("V"=>0.0437043, "E"=>0.00128242, "F"=>0.0819438),
-#     "bt"            => Dict("V"=>0.0576748, "E"=>0.0017616, "F"=>0.0515637),)
-
-function compare_errors(err1, err2) 
-    maxdiff = 0.0
-    for k1 in ["rmse", "mae"]
-        for k2 in keys(err1[k1])
-            for k3 in ["V", "E", "F"]
-                err = abs(err1[k1][k2][k3] - err2[k1][k2][k3]) / (err1[k1][k2][k3] + err2[k1][k2][k3] + sqrt(eps()))
-                maxdiff = max(maxdiff, err)
-            end 
+function test_rmse(rmse, expected)
+    for config in keys(rmse)
+        # TODO and/or warning: can't iterate over rmse because it will have virial for isolated atom
+        #for obs in keys(rmse[config])
+        for obs in keys(expected[config])
+            @test rmse[config][obs] <= expected[config][obs]
         end
-    end 
-    return maxdiff
-end 
+    end
+end
 
-acefit!(model1a, data1;
+rmse_qr = Dict(
+    "isolated_atom" => Dict("E"=>0.0, "F"=>0.0),
+    "dia"           => Dict("V"=>0.027, "E"=>0.0012, "F"=>0.024),
+    "liq"           => Dict("V"=>0.037, "E"=>0.0006, "F"=>0.16),
+    "set"           => Dict("V"=>0.057, "E"=>0.0017, "F"=>0.12),
+    "bt"            => Dict("V"=>0.08, "E"=>0.0022, "F"=>0.07),)
+
+
+acefit!(data, model; 
        data_keys...,
        weights = weights,
        solver=ACEfit.QR())
 
-acefit!(model1b, data2;
-       data_keys...,
-       weights = weights,
-       solver=ACEfit.QR())
+err = ACEpotentials.linear_errors(data, model; data_keys..., weights=weights)
 
-acefit!(model2, data2;
-       data_keys...,
-       weights = weights,
-       solver=ACEfit.QR())
-##
-
-err11 = ACEpotentials.linear_errors(data1, model1a; data_keys..., weights=weights)
-err21 = ACEpotentials.linear_errors(data2, model1b; data_keys..., weights=weights)
-err22 = ACEpotentials.linear_errors(data2, model2; data_keys..., weights=weights)
-
-@show compare_errors(err11, err22)
-@test compare_errors(err11, err22) < 0.2
-
-@warn("The model1 - data2 test fails: probably a JuLIP conversion error")
-@show compare_errors(err11, err21)
+test_rmse(err["rmse"], rmse_qr)
 
 ##
 
@@ -100,50 +65,35 @@ err22 = ACEpotentials.linear_errors(data2, model2; data_keys..., weights=weights
 addprocs(3, exeflags="--project=$(Base.active_project())")
 @everywhere using ACEpotentials
 
-acefit!(model1a, data1;
-            data_keys...,
-            weights = weights,
-            solver=ACEfit.QR())
-
-acefit!(model2, data2;
+acefit!(data, model;
             data_keys...,
             weights = weights,
             solver=ACEfit.QR())
 
 rmprocs(workers())
 
-err11d = ACEpotentials.linear_errors(data1, model1a; data_keys..., weights=weights)
-err22d = ACEpotentials.linear_errors(data2, model2; data_keys..., weights=weights)
-
-@show compare_errors(err11, err11d)
-@show compare_errors(err22, err22d)
-@test compare_errors(err11, err11d) < sqrt(eps())
-@test compare_errors(err22, err22d) < sqrt(eps())
+err_dist = ACEpotentials.linear_errors(data, model; data_keys..., weights=weights)
+test_rmse(err_dist["rmse"], rmse_qr)
 
 ##
 
-# rmse_blr = Dict(
-#          "isolated_atom" => Dict("E"=>0.0, "F"=>0.0),
-#          "set"           => Dict("V"=>0.0619346, "E"=>0.00238807, "F"=>0.121907),
-#          "dia"           => Dict("V"=>0.0333255, "E"=>0.00130242, "F"=>0.0255582),
-#          "liq"           => Dict("V"=>0.0345897, "E"=>0.000397724, "F"=>0.157461),
-#          "bt"            => Dict("V"=>0.0822944, "E"=>0.00322198, "F"=>0.062758),)
+rmse_blr = Dict(
+         "isolated_atom" => Dict("E"=>0.0, "F"=>0.0),
+         "set" => Dict("V"=>0.068, "E"=>0.0028, "F"=>0.14),
+         "dia" => Dict("V"=>0.0333255, "E"=>0.0016, "F"=>0.03),
+         "liq" => Dict("V"=>0.035, "E"=>0.00035, "F"=>0.19),
+         "bt"  => Dict("V"=>0.09, "E"=>0.0036, "F"=>0.073),)
 
-acefit!(model1a, data1;
-        data_keys...,
-        weights = weights,
-        solver = ACEfit.BLR())
 
-acefit!(model2, data2;
+acefit!(data, model;
         data_keys...,
         weights = weights,
         solver = ACEfit.BLR())        
 
-err11 = ACEpotentials.linear_errors(data1, model1a; data_keys..., weights=weights)
-err22 = ACEpotentials.linear_errors(data2, model2; data_keys..., weights=weights)
+err_blr = ACEpotentials.linear_errors(data, model; data_keys..., weights=weights)
 
-@show compare_errors(err11, err22)
-@test compare_errors(err11, err22) < 0.2
+test_rmse(err_blr["rmse"], rmse_blr)
+
 
 ##
